Refactor for more flexible lightning trainer params
