--- configs/logger/tensorboard.yaml
@@ -1,2 +1,0 @@
-_target_: lightning.pytorch.loggers.TensorBoardLogger
-save_dir: "logs/"
